+++ rl_toolkit/networks/models/actor_critic.py
@@ -1,3 +1,5 @@
+from copy import deepcopy
+
 import tensorflow as tf
 from tensorflow.keras import Model
 
@@ -43,19 +45,20 @@ class ActorCritic(Model):
         tau: float,
         init_alpha: float,
         init_noise: float,
+        merge_index: int,
         **kwargs,
     ):
         super(ActorCritic, self).__init__(**kwargs)
 
         self.gamma = tf.constant(gamma)
         self.tau = tf.constant(tau)
-        self.cum_prob = ((tf.range(n_quantiles, dtype=tf.float32) + 0.5) / n_quantiles)[
+        self.cum_prob = ((tf.range(n_quantiles, dtype=self.dtype) + 0.5) / n_quantiles)[
             tf.newaxis, tf.newaxis, :, tf.newaxis
         ]
 
         # init Lagrangian constraint
         self.log_alpha = tf.Variable(init_alpha, trainable=True, name="log_alpha")
-        self.target_entropy = tf.cast(-n_outputs, dtype=tf.float32)
+        self.target_entropy = tf.cast(-n_outputs, dtype=self.dtype)
 
         # Actor
         self.actor = Actor(
@@ -72,23 +75,49 @@ class ActorCritic(Model):
             n_quantiles=n_quantiles,
             top_quantiles_to_drop=top_quantiles_to_drop,
             n_critics=n_critics,
+            merge_index=merge_index,
         )
-
-        # Critic target
-        self.critic_target = MultiCritic(
-            units=critic_units,
-            n_quantiles=n_quantiles,
-            top_quantiles_to_drop=top_quantiles_to_drop,
-            n_critics=n_critics,
-        )
-        self._update_target(self.critic, self.critic_target, tau=1.0)
 
     def _update_target(self, net, net_targ, tau):
         for source_weight, target_weight in zip(net.variables, net_targ.variables):
             target_weight.assign(tau * source_weight + (1.0 - tau) * target_weight)
 
+    def _td_error(
+        self, next_quantiles, next_log_pi, quantiles, reward, terminal, gamma, alpha
+    ):
+        next_quantiles = tf.sort(
+            tf.reshape(next_quantiles, [next_quantiles.shape[0], -1])
+        )
+        next_quantiles = next_quantiles[:, : -self.critic_target.top_quantiles_to_drop]
+
+        # Bellman Equation
+        target_quantiles = tf.stop_gradient(
+            reward + (1.0 - terminal) * gamma * (next_quantiles - alpha * next_log_pi)
+        )
+
+        # Compute critic loss
+        pairwise_delta = (
+            target_quantiles[:, tf.newaxis, tf.newaxis, :]
+            - quantiles[:, :, :, tf.newaxis]
+        )  # batch_size, n_critics, n_quantiles, n_target_quantiles
+        loss = tf.nn.compute_average_loss(
+            tf.reduce_mean(
+                tf.math.abs(
+                    self.cum_prob - tf.cast(pairwise_delta < 0.0, dtype=self.dtype)
+                )
+                * (
+                    pairwise_delta
+                    + tf.math.softplus(-2.0 * pairwise_delta)
+                    - tf.cast(tf.math.log(2.0), pairwise_delta.dtype)
+                ),
+                axis=[1, 2, 3],
+            )
+        )
+
+        return loss
+
     def train_step(self, sample):
-        # Re-new noise matrix every update
+        # Re-new noise matrix
         self.actor.reset_noise()
 
         # Get 'Alpha'
@@ -104,47 +133,33 @@ class ActorCritic(Model):
             sample.data["next_observation"],
             with_log_prob=True,
             deterministic=False,
+            training=True,
         )
         next_quantiles = self.critic_target(
-            [sample.data["next_observation"], next_action]
+            [sample.data["next_observation"], next_action],
+            training=True,
         )
-        next_quantiles = tf.sort(
-            tf.reshape(next_quantiles, [next_quantiles.shape[0], -1])
-        )
-        next_quantiles = next_quantiles[
-            :,
-            : self.critic_target.quantiles_total
-            - self.critic_target.top_quantiles_to_drop,
-        ]
 
-        # Bellman Equation
-        target_quantiles = tf.stop_gradient(
-            tf.tanh(sample.data["reward"])
-            + (1.0 - tf.cast(sample.data["terminal"], dtype=tf.float32))
-            * self.gamma
-            * (next_quantiles - alpha * next_log_pi)
-        )
+        # Reward
+        ext_reward = tf.cast(sample.data["ext_reward"], dtype=self.dtype)
+        terminal = tf.cast(sample.data["terminal"], dtype=self.dtype)
 
         with tf.GradientTape() as tape:
-            quantiles = self.critic([sample.data["observation"], sample.data["action"]])
-
-            # Compute critic loss
-            pairwise_delta = (
-                target_quantiles[:, tf.newaxis, tf.newaxis, :]
-                - quantiles[:, :, :, tf.newaxis]
-            )  # batch_size, n_critics, n_quantiles, n_target_quantiles
-            critic_loss = tf.nn.compute_average_loss(
-                tf.reduce_mean(
-                    tf.math.abs(
-                        self.cum_prob - tf.cast(pairwise_delta < 0.0, dtype=tf.float32)
-                    )
-                    * (
-                        pairwise_delta
-                        + tf.math.softplus(-2.0 * pairwise_delta)
-                        - tf.math.log(2.0)
-                    ),
-                    axis=[1, 2, 3],
-                )
+            quantiles = self.critic(
+                [
+                    sample.data["observation"],
+                    sample.data["action"],
+                ],
+                training=True,
+            )
+            critic_loss = self._td_error(
+                next_quantiles,
+                next_log_pi,
+                quantiles,
+                ext_reward,
+                terminal,
+                self.gamma,
+                alpha,
             )
 
         # Compute gradients
@@ -155,7 +170,7 @@ class ActorCritic(Model):
 
         # -------------------- Update 'Actor' & 'Alpha' -------------------- #
         with tf.GradientTape(persistent=True) as tape:
-            quantiles, log_pi = self(sample.data["observation"])
+            quantiles, log_pi = self(sample.data["observation"], training=True)
 
             # Compute actor loss
             actor_loss = tf.nn.compute_average_loss(
@@ -188,23 +203,36 @@ class ActorCritic(Model):
             "actor_loss": actor_loss,
             "critic_loss": critic_loss,
             "alpha_loss": alpha_loss,
-            "quantiles": quantiles[0],  # logging only one randomly sampled transition
-            "log_alpha": self.log_alpha,
+            "next_log_pi": tf.reduce_mean(next_log_pi),
+            "alpha": alpha,
         }
 
-    def call(self, inputs, with_log_prob=True, deterministic=None):
+    def call(self, inputs, training=None, with_log_prob=True, deterministic=None):
         action, log_pi = self.actor(
-            inputs, with_log_prob=with_log_prob, deterministic=deterministic
+            inputs,
+            with_log_prob=with_log_prob,
+            deterministic=deterministic,
+            training=training,
         )
-        quantiles = self.critic([inputs, action])
+        quantiles = self.critic([inputs, action], training=training)
         return [quantiles, log_pi]
 
-    def compile(self, actor_optimizer, critic_optimizer, alpha_optimizer):
+    def compile(
+        self,
+        actor_optimizer,
+        critic_optimizer,
+        alpha_optimizer,
+    ):
         super(ActorCritic, self).compile()
         self.actor_optimizer = actor_optimizer
         self.critic_optimizer = critic_optimizer
         self.alpha_optimizer = alpha_optimizer
 
+    def build(self, input_shape):
+        super(ActorCritic, self).build(input_shape)
+        self.critic_target = deepcopy(self.critic)
+
     def summary(self):
         self.actor.summary()
         self.critic.summary()
+        super(ActorCritic, self).summary()