+++ rl_toolkit/networks/models/critic.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
 from tensorflow.keras import Model
-from tensorflow.keras.initializers import VarianceScaling
+from tensorflow.keras.initializers import Orthogonal
 from tensorflow.keras.layers import Activation, Add, Dense
 
 
@@ -17,57 +17,67 @@ class Critic(Model):
         - [Controlling Overestimation Bias with Truncated Mixture of Continuous Distributional Quantile Critics](https://arxiv.org/abs/2005.04269)
     """
 
-    def __init__(self, units: list, n_quantiles: int, **kwargs):
+    def __init__(self, units: list, n_quantiles: int, merge_index: int, **kwargs):
         super(Critic, self).__init__(**kwargs)
 
-        # 1. layer
-        self.fc_0 = Dense(
-            units=units[0],
-            activation="relu",
-            kernel_initializer=VarianceScaling(
-                distribution="uniform", mode="fan_in", scale=1.0
-            ),
-        )
+        # list of hidden layers
+        self.fc_layers = []
+
+        # prepare 'merge_index'
+        self.merge_index = (len(units) - 1) if merge_index == -1 else merge_index
+
+        for i, m in enumerate(units):
+            if i != self.merge_index:
+                self.fc_layers.append(
+                    Dense(
+                        units=m,
+                        activation="elu",
+                        kernel_initializer=Orthogonal(tf.sqrt(2.0)),
+                    )
+                )
+            else:
+                self.fc_layers.append(None)  # add empty layer insted of merge layer
 
         # 2. layer     TODO(markub3327): Transformer
-        self.fc_1 = Dense(
-            units=units[1],
-            kernel_initializer=VarianceScaling(
-                distribution="uniform", mode="fan_in", scale=1.0
-            ),
+        self.fc_state = Dense(
+            units=units[self.merge_index],
+            activation=None,
+            kernel_initializer=Orthogonal(tf.sqrt(2.0)),
         )
-        self.fc_2 = Dense(
-            units=units[1],
-            kernel_initializer=VarianceScaling(
-                distribution="uniform", mode="fan_in", scale=1.0
-            ),
+        self.fc_action = Dense(
+            units=units[self.merge_index],
+            activation=None,
+            kernel_initializer=Orthogonal(tf.sqrt(2.0)),
         )
         self.add_0 = Add()
-        self.activ_0 = Activation("relu")
+        self.activ_0 = Activation("elu")
 
         # Output layer
         self.quantiles = Dense(
             n_quantiles,
-            activation="linear",
-            kernel_initializer=VarianceScaling(
-                distribution="uniform", mode="fan_in", scale=1.0
-            ),
+            activation=None,
+            kernel_initializer=Orthogonal(0.01),
             name="quantiles",
         )
 
-    def call(self, inputs):
-        # 1. layer
-        state = self.fc_0(inputs[0])
+    def call(self, inputs, training=None):
+        # the first hidden layer
+        state = inputs[0]
+        for layer in self.fc_layers[: self.merge_index]:
+            state = layer(state, training=training)
 
-        # 2. layer
-        state = self.fc_1(state)
-        action = self.fc_2(inputs[1])
+        # the second layer
+        state = self.fc_state(state, training=training)
+        action = self.fc_action(inputs[1], training=training)  # projection layer
         x = self.add_0([state, action])
         x = self.activ_0(x)
 
+        # the third layer
+        for layer in self.fc_layers[(self.merge_index + 1) :]:
+            x = layer(x, training=training)
+
         # Output layer
-        quantiles = self.quantiles(x)
-        return quantiles
+        return self.quantiles(x, training=training)
 
 
 class MultiCritic(Model):
@@ -88,23 +98,26 @@ class MultiCritic(Model):
         n_quantiles: int,
         top_quantiles_to_drop: int,
         n_critics: int,
+        merge_index: int = -1,
         **kwargs
     ):
         super(MultiCritic, self).__init__(**kwargs)
 
         self.n_quantiles = n_quantiles
-        self.quantiles_total = n_quantiles * n_critics
         self.top_quantiles_to_drop = top_quantiles_to_drop
 
         # init critics
-        self.models = [Critic(units, n_quantiles) for _ in range(n_critics)]
+        self.models = [
+            Critic(units, n_quantiles, merge_index) for _ in range(n_critics)
+        ]
 
-    def call(self, inputs):
-        quantiles = tf.stack([model(inputs) for model in self.models], axis=1)
+    def call(self, inputs, training=None):
+        quantiles = tf.stack(
+            [model(inputs, training=training) for model in self.models], axis=1
+        )
         return quantiles
 
     def summary(self):
         for model in self.models:
             model.summary()
-
-    # TODO(markub3327):    def train_step(self, data):
+        super(MultiCritic, self).summary()